// frontend/src/client/api/generated/model/enumUserRole.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Marumie Backend
 * OpenAPI spec version: 0.1.0
 */

export type EnumUserRole = (typeof EnumUserRole)[keyof typeof EnumUserRole];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EnumUserRole = {
  admin: "admin",
  user: "user",
} as const;